chore: migrate to using TS, setup husky lint-staged and use pnpm as package manager
