--- news-rest-api/src/api/routes/news.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-
-
-const rotuer = Router();
-
-
-
-export default rotuer;
