--- news-rest-api/src/lib/constants.ts
@@ -1,6 +1,0 @@
-export const NEWS_API_URL = "https://newsapi.org/v2";
-
-
-// Will move to .env file
-export const NEWS_API_KEY = "91c21333ccb84a6892f4fe6ba21fe8b3";
-
