--- news-rest-api/src/lib/environment.ts
@@ -1,7 +1,0 @@
-import dotenv from 'dotenv';
-import dotenvExpand from 'dotenv-expand';
-
-const env = dotenv.config();
-dotenvExpand.expand(env);
-
-export { };
