+++ projects/news-rest-api/package.json
@@ -1,0 +1,53 @@
+{
+  "name": "news-rest-api",
+  "version": "1.0.0",
+  "description": "News REST API Airtribe Project",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "npx ts-node src/index.ts"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/nabhanh/task-manager-rest-api.git"
+  },
+  "keywords": [
+    "node",
+    "express",
+    "rest-api"
+  ],
+  "author": "nabhanh",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/nabhanh/airtribe-assignments/issues"
+  },
+  "homepage": "https://github.com/nabhanh/airtribe-assignments#readme",
+  "dependencies": {
+    "@paralleldrive/cuid2": "^2.2.1",
+    "axios": "^1.4.0",
+    "bcrypt-ts": "^4.0.0",
+    "dotenv": "^16.3.1",
+    "dotenv-expand": "^10.0.0",
+    "express": "^4.18.2",
+    "joi": "^17.9.2",
+    "jsonwebtoken": "^9.0.1",
+    "pino": "^8.15.0",
+    "pino-http": "^8.4.0"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/jsonwebtoken": "^9.0.2",
+    "@typescript-eslint/eslint-plugin": "^6.2.1",
+    "@typescript-eslint/parser": "^6.2.1",
+    "eslint": "^8.46.0",
+    "husky": "^8.0.0",
+    "babel-jest": "^29.6.2",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-unused-imports": "^3.0.0"
+  },
+  "lint-staged": {
+    "*.{js,jsx,ts,tsx}": [
+      "prettier --write",
+      "eslint --fix"
+    ]
+  }
+}
