+++ projects/news-rest-api/src/api/routes/news.ts
@@ -1,0 +1,5 @@
+import { Router } from 'express';
+
+const rotuer = Router();
+
+export default rotuer;
