+++ projects/news-rest-api/src/lib/environment.ts
@@ -1,0 +1,7 @@
+import dotenv from 'dotenv';
+import dotenvExpand from 'dotenv-expand';
+
+const env = dotenv.config();
+dotenvExpand.expand(env);
+
+export {};
