+++ projects/news-rest-api/src/types/types.d.ts
@@ -1,0 +1,8 @@
+export type User = {
+  id: number;
+  name: string;
+  email: string;
+  password: string;
+  createdAt: string;
+  updatedAt: string;
+};
