+++ projects/task-manager-rest-api/src/constants.ts
@@ -1,0 +1,5 @@
+export enum TaskPriority {
+  Low = 0,
+  Medium = 1,
+  High = 2
+}
