--- task-manager-rest-api/constants.js
@@ -1,5 +1,0 @@
-export const Priority = Object.freeze({
-  LOW: 0,
-  NORMAL: 1,
-  HIGH: 2
-});
