--- task-manager-rest-api/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "task-manager-rest-api",
-  "version": "1.0.0",
-  "description": "Task Manager REST API Airtribe Project",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node --experimental-json-modules index.js"
-  },
-  "type": "module",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/nabhanh/task-manager-rest-api.git"
-  },
-  "keywords": [
-    "node",
-    "express",
-    "rest-api"
-  ],
-  "author": "nabhanh",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/nabhanh/airtribe-assignments/issues"
-  },
-  "homepage": "https://github.com/nabhanh/airtribe-assignments#readme",
-  "dependencies": {
-    "express": "^4.18.2",
-    "joi": "^17.9.2"
-  }
-}
